Rename golinter to testlinter, to avoid confusion with "golint"
