+++ tests/util/checker/testlinter/e2etest_lint_test.go
@@ -19,36 +19,31 @@ import (
 	"testing"
 
 	"istio.io/istio/tests/util/checker"
-	"istio.io/istio/tests/util/checker/golinter"
-	"istio.io/istio/tests/util/checker/golinter/rules"
+	"istio.io/istio/tests/util/checker/testlinter/rules"
 )
 
-func TestIntegTestSkipByIssueRule(t *testing.T) {
+func TestE2eTestSkipByIssueRule(t *testing.T) {
 	clearLintRulesList()
-	golinter.LintRulesList[golinter.IntegTest] = []checker.Rule{rules.NewSkipByIssue()}
+	LintRulesList[E2eTest] = []checker.Rule{rules.NewSkipByIssue()}
 
 	rpts, _ := getReport([]string{"testdata/"})
 	expectedRpts := []string{
-		getAbsPath("testdata/integration/integtest_test.go") + ":11:2:Only t.Skip() is allowed and t.Skip() should contain an url to GitHub issue. (skip_issue)",
-		getAbsPath("testdata/integtest_integ_test.go") + ":11:2:Only t.Skip() is allowed and t.Skip() should contain an url to GitHub issue. (skip_issue)"}
+		getAbsPath("testdata/e2e/e2e_test.go") + ":11:2:Only t.Skip() is allowed and t.Skip() should contain an url to GitHub issue. (skip_issue)",
+	}
 
 	if !reflect.DeepEqual(rpts, expectedRpts) {
 		t.Errorf("lint reports don't match\nReceived: %v\nExpected: %v", rpts, expectedRpts)
 	}
 }
 
-func TestIntegTestSkipByShortRule(t *testing.T) {
+func TestE2eTestSkipByShortRule(t *testing.T) {
 	clearLintRulesList()
-	golinter.LintRulesList[golinter.IntegTest] = []checker.Rule{rules.NewSkipByShort()}
+	LintRulesList[E2eTest] = []checker.Rule{rules.NewSkipByShort()}
 
 	rpts, _ := getReport([]string{"testdata/"})
-	expectedRpts := []string{getAbsPath("testdata/integration/integtest_test.go") +
+	expectedRpts := []string{getAbsPath("testdata/e2e/e2e_test.go") +
 		":10:1:Missing either 'if testing.Short() { t.Skip() }' or 'if !testing.Short() {}' (short_skip)",
-		getAbsPath("testdata/integration/integtest_test.go") +
-			":25:1:Missing either 'if testing.Short() { t.Skip() }' or 'if !testing.Short() {}' (short_skip)",
-		getAbsPath("testdata/integtest_integ_test.go") +
-			":10:1:Missing either 'if testing.Short() { t.Skip() }' or 'if !testing.Short() {}' (short_skip)",
-		getAbsPath("testdata/integtest_integ_test.go") +
+		getAbsPath("testdata/e2e/e2e_test.go") +
 			":25:1:Missing either 'if testing.Short() { t.Skip() }' or 'if !testing.Short() {}' (short_skip)"}
 
 	if !reflect.DeepEqual(rpts, expectedRpts) {
