+++ tests/util/checker/testlinter/rules/no_goroutine.go
