+++ tests/util/checker/testlinter/rules/no_short.go
