+++ tests/util/checker/testlinter/rules/short_skip.go
