+++ tests/util/checker/testlinter/rules/skip_issue.go
