+++ tests/util/checker/testlinter/rules_matcher.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package golinter
+package main
 
 import (
 	"os"
